requirements.txt
sqlalchemy
alembic
sqlite-utils
sqlalchemy-utils
fastapi
PYYAML
uvicorn